A Python-based collection of mini-projects including a GUI Twitter Sentiment Analyzer and a Simple Calculator.
A set of mini-projects in Python that consist of a GUI Twitter Sentiment Analyzer and a Simple Calculator.
•	Project 1: Twitter Sentiment Analysis (with GUI)
This project classifies tweet sentiments (Positive, Negative, Neutral) based on NLP. It is a graphical user interface developed using “Tkinter”, and employs “NLTK” and “Scikit-learn” for natural language processing and classification.
 How to Run:
1.	Refer to the following document: “STEPS TO RUN THE CODE.docx”.
-	Verify the libraries required and the step-by-step procedure.
2.	Finish the installation and setup process.
3.	Next, click and run the file: “RUN THE CODE-(click this).py”

•	Project 2: Simple Calculator (with GUI)
A pastel-colored, light calculator constructed using “Tkinter”. It supports simple arithmetic and has a simple user interface.
How to Run
1.	Just click and run the file: “SIMPLE CALCULATOR (PROJECT-2).py”.